Move ResolveVariable into excellent and move scanner stuff into scanner.go
