+++ excellent/types/resolver.go
@@ -3,8 +3,6 @@ package types
 import (
 	"fmt"
 	"reflect"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/nyaruka/goflow/utils"
@@ -50,121 +48,6 @@ func IsNil(v interface{}) bool {
 	}
 
 	return false
-}
-
-// ResolveVariable will resolve the passed in string variable given in dot notation and return
-// the value as defined by the VariableResolver passed in.
-//
-// Example syntaxes:
-//      foo.bar.0  - 0th element of bar slice within foo, could also be "0" key in bar map within foo
-//      foo.bar[0] - same as above
-func ResolveVariable(env utils.Environment, variable interface{}, key string) interface{} {
-	var err error
-
-	err, isErr := variable.(error)
-	if isErr {
-		return err
-	}
-
-	// self referencing
-	if key == "" {
-		return variable
-	}
-
-	// strip leading '.'
-	if key[0] == '.' {
-		key = key[1:]
-	}
-
-	rest := key
-	for rest != "" {
-		key, rest = popNextVariable(rest)
-
-		if IsNil(variable) {
-			return fmt.Errorf("can't resolve key '%s' of nil", key)
-		}
-
-		// is our key numeric?
-		index, err := strconv.Atoi(key)
-		if err == nil {
-			indexable, isIndexable := variable.(Indexable)
-			if isIndexable {
-				if index >= indexable.Length() || index < -indexable.Length() {
-					return fmt.Errorf("index %d out of range for %d items", index, indexable.Length())
-				}
-				if index < 0 {
-					index += indexable.Length()
-				}
-				variable = indexable.Index(index)
-				continue
-			}
-		}
-
-		resolver, isResolver := variable.(Resolvable)
-
-		// look it up in our resolver
-		if isResolver {
-			variable = resolver.Resolve(key)
-
-			err, isErr := variable.(error)
-			if isErr {
-				return err
-			}
-
-		} else {
-			return fmt.Errorf("can't resolve key '%s' of type %s", key, reflect.TypeOf(variable))
-		}
-	}
-
-	// check what we are returning is a type that expressions understand
-	_, _, err = ToXAtom(env, variable)
-	if err != nil {
-		_, isAtomizable := variable.(Atomizable)
-		if !isAtomizable {
-			panic(fmt.Sprintf("key '%s' of resolved to usupported type %s", key, reflect.TypeOf(variable)))
-		}
-	}
-
-	return variable
-}
-
-// popNextVariable pops the next variable off our string:
-//     foo.bar.baz -> "foo", "bar.baz"
-//     foo[0].bar -> "foo", "[0].baz"
-//     foo.0.bar -> "foo", "0.baz"
-//     [0].bar -> "0", "bar"
-//     foo["my key"] -> "foo", "my key"
-func popNextVariable(input string) (string, string) {
-	var keyStart = 0
-	var keyEnd = -1
-	var restStart = -1
-
-	for i, c := range input {
-		if i == 0 && c == '[' {
-			keyStart++
-		} else if c == '[' {
-			keyEnd = i
-			restStart = i
-			break
-		} else if c == ']' {
-			keyEnd = i
-			restStart = i + 1
-			break
-		} else if c == '.' {
-			keyEnd = i
-			restStart = i + 1
-			break
-		}
-	}
-
-	if keyEnd == -1 {
-		return input, ""
-	}
-
-	key := strings.Trim(input[keyStart:keyEnd], "\"")
-	rest := input[restStart:]
-
-	return key, rest
 }
 
 type mapResolver struct {
